refactor (check_reqs.js): relocate to package root
* update tests and paths
