+++ lib/create.js
@@ -21,7 +21,7 @@ const fs = require('fs-extra');
 const path = require('path');
 const rootDir = path.resolve(__dirname, '..');
 const events = require('cordova-common').events;
-const check_reqs = require(path.join(rootDir, 'bin/templates/cordova/lib/check_reqs.js'));
+const check_reqs = require(path.join(rootDir, 'lib/check_reqs'));
 
 // exported method to create a project, returns a promise that resolves with null
 module.exports.createProject = (platform_dir, package_name, project_name, options) => {
